integration_tests/test-requirements.txt
docker-compose
pyhamcrest
pytest
requests
https://github.com/wazo-pbx/wazo-chatd-client/archive/master.zip
https://github.com/wazo-pbx/xivo-lib-rest-client/archive/master.zip
https://github.com/wazo-pbx/xivo-test-helpers/archive/master.zip

# for database tests
psycopg2-binary
sqlalchemy
https://github.com/wazo-pbx/xivo-lib-python/archive/master.zip
-e ..
